+++ app/serializers/content_block_serializer.rb
@@ -1,0 +1,15 @@
+module ClearCMS
+  class ContentBlockSerializer < ActiveModel::Serializer
+    attributes :id, :content_block_type
+    # attribute :type, :key => :content_type
+
+    def id
+      object._id.to_s
+    end
+
+    def content_block_type
+      object.type.to_s
+    end
+
+  end
+end
